Styles navbar
