+++ src/components/CategoryFilter.js
@@ -5,19 +5,16 @@ function CategoryFilter({ handleFilter }) {
   return (
     <tr>
       <td>
-        <label htmlFor="categoryFIlter">
-          category filter:
-          <select onChange={handleFilter} name="categoryFilter" id="">
-            <option value="All">All</option>
-            <option value="Action">Action</option>
-            <option value="Biography">Biography</option>
-            <option value="Kids">Kids</option>
-            <option value="Learning">Learning</option>
-            <option value="History">History</option>
-            <option value="Horror">Horror</option>
-            <option value="Sci-fi">Sci-fi</option>
-          </select>
-        </label>
+        <select className="CATEGORIES" onChange={handleFilter} name="categoryFilter" id="">
+          <option value="All">All</option>
+          <option value="Action">Action</option>
+          <option value="Biography">Biography</option>
+          <option value="Kids">Kids</option>
+          <option value="Learning">Learning</option>
+          <option value="History">History</option>
+          <option value="Horror">Horror</option>
+          <option value="Sci-fi">Sci-fi</option>
+        </select>
       </td>
     </tr>
   );
